// src/common/models/db.ts
namespace accountFields {
    export type id = number;
    export type nickName = string;
    export type password = string;
    export type email = string;
    export type phone = string | null;
    export type avatar = string;
    export type realName = string | null;
    export type identityNumber = string | null;
    export type age = number | null;
    export type company = string | null;
    export type position = string | null;
    export type personalProfile = string | null;
    export type sex = string | null;
    export type birthday = number | null;
    export type education = string | null;
    export type profession = string | null;
    export type createAt = number;
    export type status = number;

}

export interface account {
    id: accountFields.id;
    nickName: accountFields.nickName;
    password: accountFields.password;
    email: accountFields.email;
    phone: accountFields.phone;
    avatar: accountFields.avatar;
    realName: accountFields.realName;
    identityNumber: accountFields.identityNumber;
    age: accountFields.age;
    company: accountFields.company;
    position: accountFields.position;
    personalProfile: accountFields.personalProfile;
    sex: accountFields.sex;
    birthday: accountFields.birthday;
    education: accountFields.education;
    profession: accountFields.profession;
    createAt: accountFields.createAt;
    status: accountFields.status;

}
// ...
    export type type = number;